# go/model_question_answer.go
/*
 * LBA OpenUAV Questions
 *
 * Questions and answers of LBA OpenUAV for preparing to the exam.
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type QuestionAnswer struct {

	Id int32 `json:"id,omitempty"`

	Text string `json:"text,omitempty"`

	GuessedAsCorrect *bool `json:"guessedAsCorrect,omitempty"`

	IsCorrect *bool `json:"isCorrect,omitempty"`
}

// AssertQuestionAnswerRequired checks if the required fields are not zero-ed
func AssertQuestionAnswerRequired(obj QuestionAnswer) error {
	return nil
}

// AssertRecurseQuestionAnswerRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of QuestionAnswer (e.g. [][]QuestionAnswer), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseQuestionAnswerRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aQuestionAnswer, ok := obj.(QuestionAnswer)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertQuestionAnswerRequired(aQuestionAnswer)
	})
}